requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
Django==3.1.2
django-appconf==1.0.4
django-filter==2.4.0
django-mysql==3.10.0
django-sorting-bootstrap==2.5.0
django-ttag==3.0
django-userforeignkey==0.3.0
fontawesome-free==5.15.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
Pillow==8.0.1
protobuf==3.13.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
